requirements.txt
customtkinter==5.2.2
pyotp==2.9.0
pywinauto==0.6.9
bcrypt==5.0.0
cryptography==46.0.1
argon2-cffi==25.1.0
ntplib==0.4.0
psutil==7.1.3
pywin32==311
requests==2.32.5
packaging==25.0
pillow==11.3.0
pystray==0.19.5
pyinstaller==6.16.0
ruff==0.9.7
